determine the url of the nutils log in cli.call
Requires the property `log_url_prefix` to be set (in the configuration file).
The property defines the url corresponding to the `rootoutdir` property.  If
the public html directory is served in the usual way and the `rootoutdir`
property points to `'~/public_html'` (the default) the log url prefix should be
`'https://example.com/~username'`.