replace shape argument of evaluable.Unravel
The `evaluable.Unravel` class currently accepts a shape as 2-tuple. For the to
be created equivalent in the `function` module, it is convenient to pass the
shape to `evaluable.Unravel` as two separate arguments. This patch applies this
change to the class and the `_unravel` simplification method, leaving
`evaluable.unravel` unchanged.